repository and service changes

+++ src/main/java/com/bookstore/bookstore/repository/CartRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartRepository extends MongoRepository<Cart, Long> {
+public interface CartRepository extends MongoRepository<Cart, Integer> {
 }
